+++ R/utils.R
@@ -114,6 +114,8 @@ detect_repos <- function(repos = getOption("repos")) {
 #' @inheritParams cff_create
 #' @noRd
 fuzzy_keys <- function(keys) {
+  nm <- names(keys)
+  names(keys) <- gsub("_", "-", nm, fixed = TRUE)
   valid_keys <- cff_schema_keys()
 
   names <- names(keys)
@@ -137,20 +139,23 @@ fuzzy_keys <- function(keys) {
       keys_match,
       function(x) {
         if (length(x) == 0) {
-          return("No match")
+          return("No match, removing.")
         }
         return(x[1])
       }
     ))
 
     # Message
-    ll <- paste0("{.dt ", names_fuzzy, "}{.dl ", keys_match, "}\n",
-      collapse = ""
-    )
-    cli::cli_alert_warning(
-      paste0("Found misspelled keys. Trying to map:\n", ll)
+    ll <- paste0("{.dt ", names_fuzzy, "}{.dl ", keys_match, "}")
+
+    bullets <- rep("v", length(ll))
+    bullets[keys_match == "No match, removing."] <- "x"
+    names(ll) <- bullets
+    cli::cli_alert_info(
+      paste0("Found misspelled keys. Trying to map:")
     )
 
+    cli::cli_bullets(ll)
     # Modify names
     names[!is_valid_key] <- keys_match
   }